Gitlab/python-bms-viewer2/requirements.txt
setuptools==75.1.0
fonttools==4.39.3
matplotlib==3.7.1
numpy==2.1.1
pyparsing==3.0.9
PyQt5==5.15.11
python-can==4.4.2
python-dateutil==2.8.2
pytest-qt==4.4.0
